requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.2.0
langdetect==1.0.9
supabase==2.4.5
fpdf2==2.7.8
python-docx==1.1.2
